Reorganize CLI into package and redistribute tests for maintainability
Split monolithic cli.py (418 lines) into cli/ package with per-command
modules. Redistribute test_coverage_gaps.py (1196 lines) into proper
domain-specific test files. Extract shared test helpers into tests/helpers.py.
